feat: add end-to-end support helpers and tests
- Introduced support modules for end-to-end testing in Vajra, including HTTP, process, and startup helpers.
- Created a new `support.rb` file to include all helper modules for easier access in tests.
- Added comprehensive HTTP helper methods for parsing responses and handling requests.
- Implemented process management helpers to handle process lifecycle during tests.
- Developed startup helpers to manage and validate server startup scenarios.
- Removed the existing `vajra_spec.rb` file and replaced it with modularized tests for better organization and maintainability.
- Added new specs for native extension loading, header generation, public API, and version checking.
- Ensured all new features are covered by tests to maintain code quality and reliability.

closes: #33
closes: #30
