+++ gems/vajra/spec/vajra/native_extension_spec.rb
@@ -5,47 +5,31 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-RSpec.describe Vajra do
+RSpec.describe Vajra::NativeExtension do
   let(:failing_loader) do
     proc { |_path| raise LoadError, 'cannot load such file -- vajra/vajra' }
   end
 
-  it 'has a version number' do
-    expect(Vajra::VERSION).not_to be_nil
-  end
-
   it 'loads the native extension through the canonical require path' do
-    expect(
-      [described_class.respond_to?(:start), described_class.respond_to?(:stop)]
-    ).to eq([true, true])
-  end
-
-  it 'builds a versioned ASCII header' do
-    expect(described_class.header).to include('__      __')
-    expect(described_class.header).to include("v#{described_class::VERSION}")
-  end
-
-  it 'keeps the native start entrypoint public' do
-    expect(described_class).to respond_to(:start)
-    expect(described_class.singleton_methods).not_to include(:native_start)
+    expect([Vajra.respond_to?(:start), Vajra.respond_to?(:stop)]).to eq([true, true])
   end
 
   it 'raises an actionable error when the native extension cannot be loaded' do
-    expect { described_class::NativeExtension.load!(loader: failing_loader) }.to raise_error(
+    expect { described_class.load!(loader: failing_loader) }.to raise_error(
       LoadError,
       /bundle exec rake compile/
     )
   end
 
   it 'preserves the original load error backtrace' do
-    failing_loader = proc do |_path|
+    backtrace_loader = proc do |_path|
       raise LoadError, 'cannot load such file -- vajra/vajra'
     rescue LoadError => e
       raise e, e.message, ['/tmp/native_extension.rb:12']
     end
 
     begin
-      described_class::NativeExtension.load!(loader: failing_loader)
+      described_class.load!(loader: backtrace_loader)
     rescue LoadError => e
       expect(e.backtrace.first).to eq('/tmp/native_extension.rb:12')
     end
